+++ CHANGES.md
@@ -4,7 +4,9 @@
 * Changing view angle with the mouse wheel now displays change on screen for a moment.
 	* Allow to set which aspect of view angle to control, i.e. horizontal view angle, vertical view angle or focal length in the addon preferences.
 	* Works for cameras, too.
-* Can now customize mouse strafe speed on a per-blend-file basis by using Alt + Mousewheel.
+* Changing view angle with mouse wheel is now on Alt + mouse wheel to be in line with Alt reducing pan sensitivity being an interest in high zoom situations.
+	* Using Alt to lower pan speed will now lower the pan angle further depending on zoom level.
+	* Just mouse wheel can now change mouse strafe speed on a per-blend-file basis (if set instead of moving the camera in addon preferences).
 * Can now configure up to 7 mouse strafe speed multipliers which can be cycled through by pressing G or Shift+G.
 * Holding the Relocate Pivot key to disable automatic pivot relocation will no longer cause a pivot relocation.
 * Fixed toggling automic pivot relocation by holding the Relocate Pivot key not being saved in the addon preferences when closing Blender.